Add `/discover` path (#186)
Currently `gumroad.com/discover` redirects to `gumroad.com` via a
Cloudflare rule. We will need the app to serve `/discover` in order to
remove the rule.

This change is quite safe on its own because re-enabling the Cloudfare
rule effectively makes this a no-op. But this has to be rolled out
separately to avoid breaking discovery.