Fixed browser theme dependent text on Small Bets page (#516)
# Fixed Small Bets page gray text displayed incorrectly due to how
tailwind text-muted works.

## Bug Description

When the user enters the page https://gumroad.com/small-bets using a
browser with darkmode enabled, the gray text on both CTAs gets really
light and unreadable.


![obraz](https://github.com/user-attachments/assets/d9ffe994-ad16-455b-ac57-0bb8ca3a8b6d)

![obraz](https://github.com/user-attachments/assets/fba62809-cc3a-421f-a193-2668ff430e45)

## Root Cause Analysis

The text in the original version had a tailwind class "text-muted"
assigned to it. Under the hood it applies following styling to the gray
text:

`color: rgb(var(--color)/0.5);`

The thing is that --color variable is subject to lightmode darkmode
changes on user browsers. Since Gumroad doesn't have a darkmode version,
it is safe to just hardcode the color to ensure it works on both themes.

## Changes

Two lines changed in small_bets.html.erb in order to explicitly declare
gray text color.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Style**
* Updated text color styling for invitation messages on the Small Bets
page for improved visual clarity.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Laura García Diéguez <56547676+laugardie@users.noreply.github.com>
Co-authored-by: Kyle <github@kylezh.com>