Consistently authorize Helper API endpoints (#2098)
Fixes a security issue where the user tools controller was missing token
authorization, and adds authorization specs to all Helper endpoints to
catch any similar issues in future.

* Made tools token authorization the default, and added overrides for
the couple of endpoints that use HMAC instead
* Added shared examples for the authorization, and switched all specs to
include them instead of duplicating auth testing

Note this is currently patched in production; this PR is the long-term
solution.